added merging and a cooldown to prevent spamminn
the merging feature was added, which was the main feature for the game. also added a cooldown on dropping a sport ball. This was because you could spam it and then if two of the same ball were stuck together, the game would think that there were repeating and keep giving you points.